server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Nickolai Tchesnokov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.7"
openai = "^0.26.4"
selenium = "^4.8.0"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
webdriver-manager = "^3.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
